SASL

+++ random.cpp
@@ -101,8 +101,7 @@ void srand(int a, int b, int c)
 
 int rand()
 {
-
-	return abs(xsrand.rand());
+	return abs((int)xsrand.rand());
 }
 
 /**
